+++ src/main/java/com/global/api/services/HostedService.java
@@ -2,6 +2,7 @@ package com.global.api.services;
 
 import com.global.api.ServicesContainer;
 import com.global.api.builders.AuthorizationBuilder;
+import com.global.api.entities.AlternativePaymentResponse;
 import com.global.api.entities.Transaction;
 import com.global.api.entities.enums.PaymentMethodType;
 import com.global.api.entities.enums.TransactionType;
@@ -9,11 +10,14 @@ import com.global.api.entities.exceptions.ApiException;
 import com.global.api.paymentMethods.TransactionReference;
 import com.global.api.serviceConfigs.GpEcomConfig;
 import com.global.api.utils.GenerationUtils;
+import com.global.api.utils.IRequestEncoder;
 import com.global.api.utils.JsonDoc;
 import com.global.api.utils.JsonEncoders;
 
 import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Map;
 
 public class HostedService {
     GpEcomConfig _config;
@@ -59,6 +63,10 @@ public class HostedService {
 
     public Transaction parseResponse(String json, boolean encoded, String configName) throws ApiException {
         JsonDoc response = JsonDoc.parse(json, encoded ? JsonEncoders.base64Encoder() : null);
+        String merchantResponseUrl = response.getStringOrNull("MERCHANT_RESPONSE_URL");
+        if (merchantResponseUrl == null) {
+            response = mapTransactionStatusResponse(response, encoded ? JsonEncoders.base64Encoder() : null);
+        }
 
         String timestamp = response.getString("TIMESTAMP");
         String merchantId = response.getString("MERCHANT_ID");
@@ -66,10 +74,26 @@ public class HostedService {
         String result = response.getString("RESULT");
         String message = response.getString("MESSAGE");
         String transactionId = response.getString("PASREF");
-        String authCode = response.getString("AUTHCODE");
+        String authCode = response.getStringOrNull("AUTHCODE");
+        if (authCode == null) {
+            authCode = "";
+        }
+        String paymentMethod = response.getStringOrNull("PAYMENTMETHOD");
+        if (paymentMethod == null) {
+            paymentMethod = "";
+        }
 
         String sha1Hash = response.getString("SHA1HASH");
-        String hash = GenerationUtils.generateHash(_config.getSharedSecret(), timestamp, merchantId, orderId, result, message, transactionId, authCode);
+        String hash = GenerationUtils.generateHash(
+                _config.getSharedSecret(),
+                timestamp,
+                merchantId,
+                orderId,
+                result,
+                message,
+                transactionId,
+                merchantResponseUrl != null ? authCode : paymentMethod);
+
         if (!hash.equals(sha1Hash))
             throw new ApiException("Incorrect hash. Please check your code and the Developers Documentation.");
 
@@ -88,15 +112,53 @@ public class HostedService {
 
         Transaction trans = new Transaction();
         trans.setAuthorizedAmount(response.getDecimal("AMOUNT"));
-        trans.setAutoSettleFlag(response.getString("AUTO_SETTLE_FLAG"));
         trans.setCvnResponseCode(response.getString("CVNRESULT"));
         trans.setResponseCode(result);
         trans.setResponseMessage(message);
         trans.setAvsResponseCode(response.getString("AVSPOSTCODERESULT"));
-        trans.setTimestamp(timestamp);
         trans.setTransactionReference(ref);
+        trans.setAutoSettleFlag(response.getString("AUTO_SETTLE_FLAG"));
+        trans.setTimestamp(timestamp);
+
+        if (response.getStringOrNull("PAYMENTMETHOD") != null) {
+            AlternativePaymentResponse apm = new AlternativePaymentResponse();
+            apm.setCountry(response.getStringOrNull("COUNTRY"));
+            apm.setProviderName(response.getStringOrNull("PAYMENTMETHOD"));
+            apm.setPaymentStatus(response.getStringOrNull("TRANSACTION_STATUS"));
+            apm.setReasonCode(response.getStringOrNull("PAYMENT_PURPOSE"));
+            apm.setAccountHolderName(response.getStringOrNull("ACCOUNT_HOLDER_NAME"));
+            trans.setAlternativePaymentResponse(apm);
+        }
+
         trans.setResponseValues(rvalues);
 
         return trans;
     }
+
+    private JsonDoc mapTransactionStatusResponse(JsonDoc response, IRequestEncoder iRequestEncoder) {
+
+        JsonDoc newResponse = new JsonDoc(iRequestEncoder);
+
+        newResponse.set("ACCOUNT_HOLDER_NAME", response.getStringOrNull("accountholdername"));
+        newResponse.set("ACCOUNT_NUMBER", response.getStringOrNull("accountnumber"));
+        newResponse.set("TIMESTAMP", response.getStringOrNull("timestamp"));
+        newResponse.set("MERCHANT_ID", response.getStringOrNull("merchantid"));
+        newResponse.set("BANK_CODE", response.getStringOrNull("bankcode"));
+        newResponse.set("BANK_NAME", response.getStringOrNull("bankname"));
+        newResponse.set("HPP_CUSTOMER_BIC", response.getStringOrNull("bic"));
+        newResponse.set("COUNTRY", response.getStringOrNull("country"));
+        newResponse.set("HPP_CUSTOMER_EMAIL", response.getStringOrNull("customeremail"));
+        newResponse.set("TRANSACTION_STATUS", response.getStringOrNull("fundsstatus"));
+        newResponse.set("IBAN", response.getStringOrNull("iban"));
+        newResponse.set("MESSAGE", response.getStringOrNull("message"));
+        newResponse.set("ORDER_ID", response.getStringOrNull("orderid"));
+        newResponse.set("PASREF", response.getStringOrNull("pasref"));
+        newResponse.set("PAYMENTMETHOD", response.getStringOrNull("paymentmethod"));
+        newResponse.set("PAYMENT_PURPOSE", response.getStringOrNull("paymentpurpose"));
+        newResponse.set("RESULT", response.getStringOrNull("result"));
+        newResponse.set("SHA1HASH", response.getStringOrNull("sha1hash"));
+
+        return newResponse;
+    }
+
 }